+++ src/UplandLocalData/Tables/UPL.OptimizationRun.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [UPL].[OptimizationRun]
 (
 	[Id]                INT IDENTITY(1,1) NOT NULL,
-	[DiscordUserId]     BIGINT            NOT NULL,
+	[DiscordUserId]     DECIMAL(20,0)     NOT NULL,
 	[RequestedDateTime] DATETIME          NOT NULL,
 	[Filename]          VARCHAR(200)      NOT NULL,
 	[Status]            VARCHAR(20)       NOT NULL,
