+++ src/UplandLocalData/Tables/UPL.RegisteredUser.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [UPL].[RegisteredUser]
 (
 	[Id]                INT IDENTITY(1,1) NOT NULL,
-	[DiscordUserId]     BIGINT            NOT NULL,
+	[DiscordUserId]     DECIMAL(20,0)     NOT NULL,
 	[DiscordUsername]   VARCHAR(200)      NOT NULL,
 	[UplandUsername]    VARCHAR(200)              ,
 	[RunCount]          INT               NOT NULL,
